Moved filedepot code to filedepot module; make a test fixture for filedepot; properly implement filedepot protocol in fixture
